Fix GGUF string lengths to uint64 for v2+ compatibility
String lengths in GGUF v2/v3 are uint64, not uint32. Updated GgufReader
and all synthetic test data. All 18 tests now pass including integration
tests against real Qwen 3.5 0.8B Q8_0 model.

+++ src/Daisi.Llama/Gguf/GgufReader.cs
@@ -71,7 +71,8 @@ public sealed class GgufReader
 
     private string ReadString()
     {
-        var length = _br.ReadUInt32();
+        // GGUF v2+ uses uint64 for string lengths
+        var length = _br.ReadUInt64();
         var bytes = _br.ReadBytes((int)length);
         return Encoding.UTF8.GetString(bytes);
     }
