+++ tests/Daisi.Llama.Tests/Gguf/GgufMetadataTests.cs
@@ -5,22 +5,22 @@ namespace Daisi.Llama.Tests.Gguf;
 /// <summary>
 /// Tests for GGUF metadata key-value parsing.
 /// Uses synthetic byte streams for deterministic testing.
+/// String lengths are uint64 in GGUF v2+.
 /// </summary>
 public class GgufMetadataTests
 {
     [Fact]
     public void ReadMetadataKv_StringValue_ParsesCorrectly()
     {
-        // Build a single KV pair: key="general.architecture", type=STRING, value="qwen35"
         using var ms = new MemoryStream();
         using var bw = new BinaryWriter(ms);
 
         var key = "general.architecture"u8;
-        bw.Write((uint)key.Length);
+        bw.Write((ulong)key.Length);
         bw.Write(key);
-        bw.Write((uint)GgufMetadataValueType.String); // type
+        bw.Write((uint)GgufMetadataValueType.String);
         var val = "qwen35"u8;
-        bw.Write((uint)val.Length);
+        bw.Write((ulong)val.Length);
         bw.Write(val);
 
         ms.Position = 0;
@@ -39,7 +39,7 @@ public class GgufMetadataTests
         using var bw = new BinaryWriter(ms);
 
         var key = "general.alignment"u8;
-        bw.Write((uint)key.Length);
+        bw.Write((ulong)key.Length);
         bw.Write(key);
         bw.Write((uint)GgufMetadataValueType.Uint32);
         bw.Write((uint)32);
@@ -59,7 +59,7 @@ public class GgufMetadataTests
         using var bw = new BinaryWriter(ms);
 
         var key = "tokenizer.score"u8;
-        bw.Write((uint)key.Length);
+        bw.Write((ulong)key.Length);
         bw.Write(key);
         bw.Write((uint)GgufMetadataValueType.Float32);
         bw.Write(0.5f);
@@ -79,7 +79,7 @@ public class GgufMetadataTests
         using var bw = new BinaryWriter(ms);
 
         var key = "general.file_type"u8;
-        bw.Write((uint)key.Length);
+        bw.Write((ulong)key.Length);
         bw.Write(key);
         bw.Write((uint)GgufMetadataValueType.Bool);
         bw.Write((byte)1);
@@ -98,7 +98,7 @@ public class GgufMetadataTests
         using var bw = new BinaryWriter(ms);
 
         var key = "tokenizer.tokens"u8;
-        bw.Write((uint)key.Length);
+        bw.Write((ulong)key.Length);
         bw.Write(key);
         bw.Write((uint)GgufMetadataValueType.Array);
         // Array: element_type(uint32) + count(uint64) + elements
@@ -107,7 +107,7 @@ public class GgufMetadataTests
         foreach (var token in new[] { "hello", "world", "test" })
         {
             var bytes = System.Text.Encoding.UTF8.GetBytes(token);
-            bw.Write((uint)bytes.Length);
+            bw.Write((ulong)bytes.Length);
             bw.Write(bytes);
         }
 
@@ -130,7 +130,7 @@ public class GgufMetadataTests
         using var bw = new BinaryWriter(ms);
 
         var key = "llm.context_length"u8;
-        bw.Write((uint)key.Length);
+        bw.Write((ulong)key.Length);
         bw.Write(key);
         bw.Write((uint)GgufMetadataValueType.Uint64);
         bw.Write((ulong)262144);
